refactor: drop "pkg" directory from "route" module
Related to #44.
