+++ controlplane/pkg/yncp/cfg.go
@@ -8,7 +8,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/yanet-platform/yanet2/controlplane/internal/pkg/gateway"
-	"github.com/yanet-platform/yanet2/controlplane/modules/route/pkg/route"
+	"github.com/yanet-platform/yanet2/controlplane/modules/route"
 )
 
 type Config struct {
